create other necessary fields and methods at header files
