--- FileSender_Package/Encryption_Package/Encryption.h
@@ -1,15 +1,0 @@
-
-#ifndef SINA_SFILESENDER_ENCRYPTION_H
-#define SINA_SFILESENDER_ENCRYPTION_H
-
-
-class Encryption {
-
-public:
-    Encryption();
-    void encryptDirectory();
-    void decryptDirectory();
-};
-
-
-#endif
