+++ test/UnitTest.h
@@ -1,0 +1,14 @@
+//
+// Created by Sina Elahimanesh on 2020-06-14.
+//
+
+#ifndef CPLUSCLASS1399_UNITTEST_H
+#define CPLUSCLASS1399_UNITTEST_H
+
+
+class UnitTest {
+
+};
+
+
+#endif //CPLUSCLASS1399_UNITTEST_H
